Rename and deprecate api route handler wrappers
