+++ packages/nextjs/src/server/wrapApiHandlerWithSentry.ts
@@ -25,7 +25,7 @@ import { autoEndTransactionOnResponseEnd, finishTransaction, flushQueue } from '
  * @param parameterizedRoute The page's route, passed in via the proxy loader
  * @returns The wrapped handler
  */
-export function withSentryAPI(
+export function wrapApiWithSentry(
   maybeWrappedHandler: NextApiHandler | WrappedNextApiHandler,
   parameterizedRoute: string,
 ): WrappedNextApiHandler {
@@ -47,8 +47,14 @@ export function withSentryAPI(
     );
   }
 
+  // eslint-disable-next-line deprecation/deprecation
   return withSentry(maybeWrappedHandler, parameterizedRoute);
 }
+
+/**
+ * @deprecated Use `wrapApiWithSentry()` instead
+ */
+export const withSentryAPI = wrapApiWithSentry;
 
 /**
  * Legacy function for manually wrapping API route handlers, now used as the innards of `withSentryAPI`.
@@ -56,6 +62,8 @@ export function withSentryAPI(
  * @param origHandler The user's original API route handler
  * @param parameterizedRoute The route whose handler is being wrapped. Meant for internal use only.
  * @returns A wrapped version of the handler
+ *
+ * @deprecated Use `wrapApiWithSentry()` instead
  */
 export function withSentry(origHandler: NextApiHandler, parameterizedRoute?: string): WrappedNextApiHandler {
   // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
